# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant School Management SaaS API",
        "title": "EduAccess API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "EduAccess Support",
            "email": "support@eduaccess.id"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/teachers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of teachers. Superadmin may filter by school_id; admin_sekolah is scoped to their own school.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "List teachers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School UUID (superadmin only)",
                        "name": "school_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name, email or username",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/http.TeacherResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a user account (role=teacher) and teacher profile atomically.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Create teacher",
                "parameters": [
                    {
                        "description": "Teacher data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateTeacherRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.TeacherResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/teachers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Get teacher by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Teacher profile UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.TeacherResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates mutable fields in users and teacher_profiles.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Update teacher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Teacher profile UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Teacher fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateTeacherRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.TeacherResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Deactivate teacher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Teacher profile UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.CreateTeacherRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "username"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "berat_badan": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "birth_place": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "other"
                    ]
                },
                "golongan_darah": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "kelainan_jasmani": {
                    "type": "string"
                },
                "kewarganegaraan": {
                    "type": "string"
                },
                "kode_pos": {
                    "type": "string"
                },
                "ktp_image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 191,
                    "minLength": 1
                },
                "nik": {
                    "type": "string"
                },
                "nip": {
                    "type": "string"
                },
                "nuptk": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "pendidikan_terakhir": {
                    "type": "string"
                },
                "penyakit_kronis_yang_pernah_diderita": {
                    "type": "string"
                },
                "penyakit_yang_sering_kambuh": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "religion": {
                    "type": "string"
                },
                "rt_rw": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "tahun_lulus": {
                    "type": "string"
                },
                "tahun_masuk": {
                    "type": "string"
                },
                "tinggi_badan": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 191,
                    "minLength": 3
                }
            }
        },
...
        "http.TeacherResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "berat_badan": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "birth_place": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "golongan_darah": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "kelainan_jasmani": {
                    "type": "string"
                },
                "kewarganegaraan": {
                    "type": "string"
                },
                "kode_pos": {
                    "type": "string"
                },
                "ktp_image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "nip": {
                    "type": "string"
                },
                "nuptk": {
                    "type": "string"
                },
                "pendidikan_terakhir": {
                    "type": "string"
                },
                "penyakit_kronis_yang_pernah_diderita": {
                    "type": "string"
                },
                "penyakit_yang_sering_kambuh": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "religion": {
                    "type": "string"
                },
                "rt_rw": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "tahun_lulus": {
                    "type": "string"
                },
                "tahun_masuk": {
                    "type": "string"
                },
                "tinggi_badan": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "http.UpdateTeacherRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "berat_badan": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "birth_place": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "other"
                    ]
                },
                "golongan_darah": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "kelainan_jasmani": {
                    "type": "string"
                },
                "kewarganegaraan": {
                    "type": "string"
                },
                "kode_pos": {
                    "type": "string"
                },
                "ktp_image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 191,
                    "minLength": 1
                },
                "nik": {
                    "type": "string"
                },
                "nip": {
                    "type": "string"
                },
                "nuptk": {
                    "type": "string"
                },
                "pendidikan_terakhir": {
                    "type": "string"
                },
                "penyakit_kronis_yang_pernah_diderita": {
                    "type": "string"
                },
                "penyakit_yang_sering_kambuh": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "religion": {
                    "type": "string"
                },
                "rt_rw": {
                    "type": "string"
                },
                "tahun_lulus": {
                    "type": "string"
                },
                "tahun_masuk": {
                    "type": "string"
                },
                "tinggi_badan": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 191,
                    "minLength": 3
                }
            }
        },
...
        "response.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and a Supabase JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}